requirements.txt
amadeus
openapi-client
fastapi
uvicorn
python-dotenv
pydantic
langchain_openai
fpdf
python-docx